Roberio_04_07_11_51

+++ src/main/java/br/com/tisyserp/model/TabelaAuxiliar/Cfop.java
@@ -20,7 +20,7 @@ public class Cfop extends PanacheEntityBase  {
 	@Id
 	@Basic(optional = false)
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	public Integer cfop_id;
+	public Long cfop_id;
 	
 	@NotNull
 	@Size(max = 4)
